🔥 [REFACTOR] 메인 제외 담당 중복  ViewController 삭제
