+++ ChioskApp/ChioskApp/Sources/Controller/MainViewController.swift
@@ -12,29 +12,74 @@ import Then
 
 class MainViewController: UIViewController {
     
-    let menuListNavigator = MenuListNavigator()
-
+    
+    let menuCategoryViews = MenuCategoryView()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupUI()
-        setupConstraints()
-    }
 
-    // UI 설정
-    private func setupUI() {
-        // menuListNavigator를 하위 뷰로 추가
-        addChild(menuListNavigator)
-        view.addSubview(menuListNavigator.view)
-        menuListNavigator.didMove(toParent: self)
+        configureUI()
+        segmentChanged(menuCategoryViews.segmentControl) //SegmentedControl
     }
-
-    // 제약 조건 설정
-    private func setupConstraints() {
-        menuListNavigator.view.snp.makeConstraints { make in
-            make.edges.equalToSuperview() // 전체 화면에 꽉 차게
+   
+    //MARK: 로고, SegmentedControl, Place Holder
+    func configureUI() {
+        view.backgroundColor = .white
+        
+        // MARK: 상단 로고
+        view.addSubview(menuCategoryViews.logo)
+        menuCategoryViews.logo.snp.makeConstraints {
+            $0.top.height.equalTo(60)
+            $0.leading.trailing.equalToSuperview()
+        }
+        
+        //MARK: SegmentedControl
+        view.addSubview(menuCategoryViews.segmentControl)
+        menuCategoryViews.segmentControl.addTarget(self, action: #selector(segmentChanged(_:)), for: .valueChanged)
+        menuCategoryViews.segmentControl.snp.makeConstraints {
+            $0.top.equalTo(menuCategoryViews.logo.snp.bottom)
+            $0.leading.trailing.equalToSuperview()
+            $0.height.equalTo(58)
+        }
+        
+        menuCategoryViews.segmentControl.layer.cornerRadius = 0
+        
+        // MARK: Place Holder 뷰들 레이아웃
+        [menuCategoryViews.chickenView, menuCategoryViews.sidedishView, menuCategoryViews.drinkView, menuCategoryViews.etcView].forEach {
+            view.addSubview($0)
+            $0.snp.makeConstraints {
+                $0.width.equalToSuperview()
+                $0.height.equalTo(352)
+                $0.top.equalTo(menuCategoryViews.segmentControl.snp.bottom)
+            }
+        }
+    
+    }
+    
+    //MARK: 클릭 시 뷰 변경되는 로직
+    @objc func segmentChanged(_ sender: UISegmentedControl) {
+        [menuCategoryViews.chickenView, menuCategoryViews.sidedishView, menuCategoryViews.drinkView, menuCategoryViews.etcView].forEach {
+            $0.isHidden = true // 해당 뷰 제외 나머지는 안보이게 숨김.
+        }
+        
+        switch sender.selectedSegmentIndex {
+        case 0:
+            menuCategoryViews.chickenView.isHidden = false
+        case 1:
+            menuCategoryViews.sidedishView.isHidden = false
+        case 2:
+            menuCategoryViews.drinkView.isHidden = false
+        case 3:
+            menuCategoryViews.etcView.isHidden = false
+        default:
+            break
         }
     }
+    
+    
 }
+
+
 
 #Preview("MainViewController") {
     MainViewController()
