+++ ChioskApp/ChioskApp/Sources/View/MenuCategoryView.swift
@@ -14,8 +14,9 @@ class MenuCategoryView {
     //MARK: SegmentedControl 구현
     let segmentControl = UISegmentedControl(items: ["치킨", "사이드", "음료", "기타"]).then {
         $0.selectedSegmentIndex = 0
-        $0.backgroundColor = .systemGray6
+        $0.backgroundColor = .darkGray
         $0.selectedSegmentTintColor = .white
+        $0.setDividerImage(UIImage(), forLeftSegmentState: .normal, rightSegmentState: .normal, barMetrics: .default)//Divider(칸막이)제거
     }
     // 클릭했을 때, 뷰를 변환되게 하는 메소드. 클릭 액션 메소드. 이것에 대한 로직을 이제 함수로 구현해야함
     // 초기 뷰를 선언해야함.
